+++ history.md
@@ -44,7 +44,10 @@ node starsky-tools/build-tools/app-version-update.js
 - remove newtonsoft.json references
 
 # version 0.4.10 _(Unreleased)_ - 2021-06-??
-- [x]   (Fixed) _Back-end_ Performance change, FileIndexItem uses less memory in the application
+- [x]   (Fixed) _Back-end_ Performance change, FileIndexItem uses less memory in the application (PR #410)
+- [x]   (Fixed) _Back-end_ Change Replace to use a single database query and update to empty string (PR #412)
+- [x]   (Fixed) _Back-end_ Fix fast updating items to update the cache before update (PR #412)
+- [x]   (Change) _Back-end_ Push direct to socket when update or replace to avoid undo after a second (PR #412)
 
 # version 0.4.9 - 2021-06-17
 - [x]   (Fixed) _Front-end_ Show error when update fails in archive list (PR #391)